compiler: keep track of temp variables that contain param/const
Keep track of what const/param is loaded into a temp variable. We can
then see if the const is already loaded in a variable and simply reuse
it instead of using a new temp. We must also make sure we load constants
of different size in different temp variables.