orcmips: only enable if the DSPr2 ASE is present
The mips target uses various instructions that a part of the MIPS DSPr2
ASE which is not present on all processors. The easiest solution is to
gate the target on the presense of DSPr2 at runtime.

Unfortunately there isn't a "nice" way to detect if DSPr2 is present, so
resort to parsing /proc/cpuinfo to get the information.

https://bugzilla.gnome.org/show_bug.cgi?id=760834