orconce: Prefer the Win32 once implementation only for MSVC
When profiling the impact on the provided test case under MSVC 17.9
Preview 1, I could observe that the slowdown wrt. the C11 atomics
stem from both an excessive number of memory accesses (because
the iteration number is not defined as a constant, rather as a
heap allocated variable), but also because the branch prediction
is a guaranteed failure. This is because the codegen prefers
to jump to the most *likely* branch instead of to the most unlikely
one.

This insight can be applied with MinGW compilers on Windows,
to instead deliver a >2x speedup by marking which branches refer
to the already initialised state. Attempting to do this to
MSVC by rewriting the conditionals instead end up adding a 1.5x
slowdown, so I've enforced a condition that makes that compiler
use the Win32 based approach.

Part-of: <https://gitlab.freedesktop.org/gstreamer/orc/-/merge_requests/117>