div255w: fix off-by-one error in the implementations
The current approximation employs unsigned division by 256, which
yields an incorrect answer 50% of the time [1]. The correct
approach, if one is trying to avoid the usage of integer division, is
to instead divide by 0x800000 / 0x8081 (~254.9996), which is done by
multiplying by the divisor, extracting the upper half of each result,
and then right shifting (dividing) 7 times, in this case.

Alternatively, one can reinterpret each temporary register as 32-bit,
and shift by 23 in situ, then extract the low half of each.

This approach was validated in a Godbolt of the C emulation [2]. See
also [3] for another reference of the approach.

Closes #7

[1]: https://tttapa.github.io/Pages/Raspberry-Pi/NEON/Division.html#dividing-by-255
[2]: https://godbolt.org/z/vWPzToMxb, https://godbolt.org/z/e73s85Kh8
[3]: Van Aken, J. R. (2022). Alpha Blending with No Division Operations. arXiv preprint arXiv:2202.02864.

Part-of: <https://gitlab.freedesktop.org/gstreamer/orc/-/merge_requests/134>