powerpc: fix div255w which still used the inexact substitution
The code for this architecture used the substitution suggested in the
original bug report. That one had subtle failures on armv7a and aarch64,
but I was not able to verify if they affected PowerPC too.

This commit fixes it by reusing the mulhuw instruction implementation --
sse2 can be ported directly that way.

Fixes #71

Part-of: <https://gitlab.freedesktop.org/gstreamer/orc/-/merge_requests/195>