constant: Improve the constant handling
OrcConstant was an attempt to define constants programmatically but
it is not used on every target and is limited to 128-bits. Extend
such definition to store at most 512-bits and to splat 64-bits values.

Add methods to resolve a constant and to compare them.

Part-of: <https://gitlab.freedesktop.org/gstreamer/orc/-/merge_requests/233>