Add support for ORC_DEBUG_FATAL envvar
This environment variable can be defined with a debug level in the same way
as ORC_DEBUG. The behavior will be to abort if there is any log message
with a level equal or less than the defined. Note that if ORC_DEBUG is set,
for example, to zero, there will be no log message, so ORC_DEBUG_FATAL will
no abort in any case.

Using compiler intrinsics "__builtin_trap()" and "__debugbreak()" to to
make it easily debuggable/backtraceable.

Part-of: <https://gitlab.freedesktop.org/gstreamer/orc/-/merge_requests/246>